Move fast_thread_switching fixture to top level
